+++ omniisaacgymenvs/mujoco_envs/environments/mujoco_base_env.py
@@ -6,226 +6,14 @@ import mujoco
 import math
 import os
 
-default_cfg = {}
-default_cfg["seed"] = 42
-default_cfg['disturbances'] = {}
-default_cfg['disturbances']['use_uneven_floor'] = False
-default_cfg['disturbances']['use_sinusoidal_floor'] = False
-default_cfg['disturbances']['floor_min_freq'] = 0.25
-default_cfg['disturbances']['floor_max_freq'] = 3.0
-default_cfg['disturbances']['floor_min_offset'] = -6.0
-default_cfg['disturbances']['floor_max_offset'] = 6.0
-default_cfg['disturbances']['max_floor_force'] = 0.4
-default_cfg['disturbances']['min_floor_force'] = 0.4
-default_cfg['disturbances']['use_torque_diturbance'] = False
-default_cfg['disturbances']['use_sinusoidal_torque'] = False
-default_cfg['disturbances']['max_torque'] = 0.1
-default_cfg['disturbances']['min_torque'] = 0.1
-default_cfg['disturbances']['add_noise_on_pos'] = False
-default_cfg['disturbances']['position_noise_min'] = -0.01
-default_cfg['disturbances']['position_noise_max'] = 0.01
-default_cfg['disturbances']['add_noise_on_vel'] = False
-default_cfg['disturbances']['velocity_noise_min'] = -0.01
-default_cfg['disturbances']['velocity_noise_max'] = 0.01
-default_cfg['disturbances']['add_noise_on_heading'] = False
-default_cfg['disturbances']['heading_noise_min'] = -0.025
-default_cfg['disturbances']['heading_noise_max'] = 0.025
-default_cfg['disturbances']['add_noise_on_act'] = False
-default_cfg['disturbances']['min_action_noise'] = -0.05
-default_cfg['disturbances']['max_action_noise'] = 0.05
-default_cfg['disturbances']['num_thrusters_to_kill'] = 0
-default_cfg["max_spawn_dist"] = 4.0
-default_cfg["min_spawn_dist"] = 3.0
-default_cfg["kill_dist"] = 6.0
-
-class RandomSpawn:
-    def __init__(self, cfg):
-        self._rng = np.random.default_rng()
-        self._max_spawn_dist = cfg["max_spawn_dist"]
-        self._min_spawn_dist = cfg["min_spawn_dist"]
-        self._kill_dist = cfg["kill_dist"]
-
-    def getInitialCondition(self):
-        theta = self._rng.uniform(-np.pi, np.pi, 1)
-        r = self._rng.uniform(self._min_spawn_dist, self._max_spawn_dist)
-        initial_position = [np.cos(theta) * r, np.sin(theta) * r]
-        heading = self._rng.uniform(-np.pi, np.pi, 1)
-        initial_orientation = [np.cos(heading*0.5), 0, 0, np.sin(heading*0.5)]
-        return initial_position, initial_orientation
-
-class RandomKillThrusters:
-    """
-    Randomly kills thrusters."""
-
-    def __init__(self, cfg):
-        self._rng = np.random.default_rng()
-        self._num_thrusters_to_kill = cfg['disturbances']['num_thrusters_to_kill']
-        self.killed_thrusters_id = []
-
-    def generate_thruster_kills(self):
-        self.killed_thrusters_id = self._rng.choice(8, self._num_thrusters_to_kill, replace=False) #[2,3]
-        print("Killed thrusters: ", self.killed_thrusters_id)
-        
-class UnevenFloorDisturbance:
-    """
-    Creates disturbances on the platform by simulating an uneven floor."""
-
-    def __init__(self, cfg: dict) -> None:
-        # Uneven floor generation
-        self._rng = np.random.default_rng(cfg["seed"])
-        self._use_uneven_floor = cfg['disturbances']['use_uneven_floor']
-        self._use_sinusoidal_floor = cfg['disturbances']['use_sinusoidal_floor']
-        self._min_freq = cfg['disturbances']['floor_min_freq']
-        self._max_freq = cfg['disturbances']['floor_max_freq']
-        self._min_offset = cfg['disturbances']['floor_min_offset']
-        self._max_offset = cfg['disturbances']['floor_max_offset']
-        self._max_floor_force = cfg['disturbances']['max_floor_force'] 
-        self._min_floor_force = cfg['disturbances']['min_floor_force'] 
-        self._max_floor_force = math.sqrt(self._max_floor_force**2 / 2)
-        self._min_floor_force = math.sqrt(self._min_floor_force**2 / 2)
-
-        self._floor_forces = np.zeros(3, dtype=np.float32)
-        self._floor_x_freq = 0
-        self._floor_y_freq = 0
-        self._floor_x_offset = 0
-        self._floor_y_offset = 0
-
-    def generate_floor(self) -> None:
-        """
-        Generates the uneven floor."""
-        if self._use_uneven_floor:
-            if self._use_sinusoidal_floor:
-                self._floor_x_freq   = self._rng.uniform(self._min_freq, self._max_freq, 1)
-                self._floor_y_freq   = self._rng.uniform(self._min_freq, self._max_freq, 1)
-                self._floor_x_offset = self._rng.uniform(self._min_offset, self._max_offset, 1)
-                self._floor_y_offset = self._rng.uniform(self._min_offset, self._max_offset, 1)
-            else:
-                r = self._rng.uniform(self._min_floor_force, self._max_floor_force, 1)
-                theta = self._rng.uniform(0, 1, 1) * math.pi * 2
-                self._floor_forces[0] = np.cos(theta) * r
-                self._floor_forces[1] = np.sin(theta) * r
-
-    def get_floor_forces(self, root_pos: np.ndarray) -> np.ndarray:
-        """
-        Computes the floor forces for the current state of the robot."""
-        if self._use_uneven_floor:
-            if self._use_sinusoidal_floor:
-                self._floor_forces[0] = np.sin(root_pos[0] * self._floor_x_freq + self._floor_x_offset) * self._max_floor_force
-                self._floor_forces[1] = np.sin(root_pos[1] * self._floor_y_freq + self._floor_y_offset) * self._max_floor_force
-       
-        return self._floor_forces
-
-
-class TorqueDisturbance:
-    """
-    Creates disturbances on the platform by simulating a torque applied to its center."""
-
-    def __init__(self, cfg: dict) -> None:
-        self._rng = np.random.default_rng(cfg["seed"])
-        # Uneven floor generation
-        self._use_torque_disturbance = cfg['disturbances']['use_torque_diturbance']
-        self._use_sinusoidal_torque = cfg['disturbances']['use_sinusoidal_torque']
-        self._max_torque = cfg['disturbances']['max_torque']
-        self._min_torque = cfg['disturbances']['min_torque']
-
-        # use the same min/max frequencies and offsets for the floor
-        self._min_freq = cfg['disturbances']['floor_min_freq']
-        self._max_freq = cfg['disturbances']['floor_max_freq']
-        self._min_offset = cfg['disturbances']['floor_min_offset']
-        self._max_offset = cfg['disturbances']['floor_max_offset']
-
-        self._torque_forces = np.zeros(3, dtype=np.float32)
-        self._torque_freq = 0
-        self._torque_offset = 0
-
-    def generate_torque(self) -> None:
-        """
-        Generates the torque disturbance."""
-        if self._use_torque_disturbance:
-            if self._use_sinusoidal_torque:
-                #  use the same min/max frequencies and offsets for the floor
-                self._torque_freq = self._rng.uniform(self._min_freq, self._max_freq, 1)
-                self._torque_offset = self._rng.uniform(self._min_offset, self._max_offset, 1)
-            else:
-                r = self._rng.uniform(self._min_torque, self._max_torque, 1) * self._rng.choice([1,-1])
-                self._torque_forces[2] = r
-
-    def get_torque_disturbance(self, root_pos: np.ndarray) -> np.ndarray:
-        """
-        Computes the torque for the current state of the robot."""
-        if self._use_torque_disturbance:
-            if self._use_sinusoidal_torque:
-                self._torque_forces[2] = np.sin(root_pos * self._torque_freq + self._torque_offset) * self._max_torque
-
-        return self._torque_forces
-
-
-class NoisyObservations:
-    """
-    Adds noise to the observations of the robot."""
-
-    def __init__(self, cfg: dict) -> None:
-        self._rng = np.random.default_rng(seed=cfg['seed'])
-        self._add_noise_on_pos = cfg['disturbances']['add_noise_on_pos']
-        self._position_noise_min = cfg['disturbances']['position_noise_min']
-        self._position_noise_max = cfg['disturbances']['position_noise_max']
-        self._add_noise_on_vel = cfg['disturbances']['add_noise_on_vel']
-        self._velocity_noise_min = cfg['disturbances']['velocity_noise_min']
-        self._velocity_noise_max = cfg['disturbances']['velocity_noise_max']
-        self._add_noise_on_heading = cfg['disturbances']['add_noise_on_heading']
-        self._heading_noise_min = cfg['disturbances']['heading_noise_min']
-        self._heading_noise_max = cfg['disturbances']['heading_noise_max']
-    
-    def add_noise_on_pos(self, pos: np.ndarray) -> np.ndarray:
-        """
-        Adds noise to the position of the robot."""
-
-        if self._add_noise_on_pos:
-            pos += self._rng.uniform(self._position_noise_min, self._position_noise_max, pos.shape)
-        return pos
-    
-    def add_noise_on_vel(self, vel: np.ndarray) -> np.ndarray:
-        """
-        Adds noise to the velocity of the robot."""
-
-        if self._add_noise_on_vel:
-            vel += self._rng.uniform(self._velocity_noise_min, self._velocity_noise_max, vel.shape)
-        return vel
-    
-    def add_noise_on_heading(self, heading: np.ndarray) -> np.ndarray:
-        """
-        Adds noise to the heading of the robot."""
-
-        if self._add_noise_on_heading:
-            heading += self._rng.uniform(self._heading_noise_min, self._heading_noise_max, heading.shape)
-        return heading
-
-
-class NoisyActions:
-    """
-    Adds noise to the actions of the robot."""
-
-    def __init__(self, cfg: dict) -> None:
-        self._rng = np.random.default_rng(seed=cfg['seed'])
-        self._add_noise_on_act = cfg['disturbances']['add_noise_on_act']
-        self._min_action_noise = cfg['disturbances']['min_action_noise']
-        self._max_action_noise = cfg['disturbances']['max_action_noise']
-
-    def add_noise_on_act(self, act: np.ndarray) -> np.ndarray:
-        """
-        Adds noise to the actions of the robot."""
-
-        if self._add_noise_on_act:
-            act += self._rng.uniform(self._min_action_noise, self._max_action_noise, 1)
-        return act
-    
+from omniisaacgymenvs.mujoco_envs.environments.disturbances import NoisyActions, NoisyObservations, TorqueDisturbance, UnevenFloorDisturbance, RandomKillThrusters, RandomSpawn
 
 class MuJoCoFloatingPlatform:
     """
     A class for the MuJoCo Floating Platform environment."""
 
     def __init__(self, step_time:float = 0.02, duration:float = 60.0, inv_play_rate:int = 10,
-                 mass:float = 5.32, max_thrust:float = 1.0, radius:float = 0.31, cfg:dict=default_cfg) -> None:
+                 mass:float = 5.32, max_thrust:float = 1.0, radius:float = 0.31, cfg:dict=None) -> None:
         """
         Initializes the MuJoCo Floating Platform environment.
         step_time: The time between steps in the simulation.
@@ -409,31 +197,18 @@ class MuJoCoFloatingPlatform:
         self.logs["quaternion"].append(np.roll(self.data.qpos[3:].copy(),-1))
         self.logs["actions"].append(self.actions)
 
-    def updateState(self) -> Dict[str, np.ndarray]:
+    def getObs(self) -> Dict[str, np.ndarray]:
         """
-        Updates the state of the simulation."""
+        returns an up to date observation buffer."""
 
-        qpos = self.data.qpos.copy() # Copy the pose of the object.
-        pos = qpos[:2]
-        pos = self.ON.add_noise_on_pos(qpos[:2])
-        # Cast the quaternion to the yaw (roll and pitch are invariant).
-        # Compute the heading
-        siny_cosp = 2 * (qpos[3] * qpos[6] + qpos[4] * qpos[5])
-        cosy_cosp = 1 - 2 * (qpos[5] * qpos[5] + qpos[6] * qpos[6])
-        orient_z = np.arctan2(siny_cosp, cosy_cosp)
-        orient_z = self.ON.add_noise_on_heading(orient_z) 
-        siny_cosp = np.sin(orient_z)
-        cosy_cosp = np.cos(orient_z)
-        # Gets the angular and linear velocity.
-        linear_velocity = self.data.qvel[0:2].copy() # X and Y velocities.
-        linear_velocity = self.ON.add_noise_on_vel(linear_velocity)
-        angular_velocity = self.data.qvel[5].copy() # Yaw velocity.
-        angular_velocity = self.ON.add_noise_on_vel(angular_velocity)
-        # Returns the state.
-        state = {"orientation": [cosy_cosp, siny_cosp], "position": pos, "linear_velocity": linear_velocity, "angular_velocity": angular_velocity}
+        state = {}
+        state["angular_velocity"] = self.ON.add_noise_on_vel(self.data.qvel[3:6].copy())
+        state["linear_velocity"] = self.ON.add_noise_on_vel(self.data.qvel[0:3].copy())
+        state["position"] = self.ON.add_noise_on_pos(self.data.qpos[0:3].copy())
+        state["quaternion"] = self.data.qpos[3:].copy()
         return state
 
-    def runLoop(self, model, initial_position=[0,0], initial_orientation=[1,0,0,0]) -> None:
+    def runLoop(self, model, initial_position=[0,0], initial_orientation=[1,0,0,0]) -> Dict[str, np.ndarray]:
         """
         Runs the simulation loop.
         model: the agent.
@@ -442,7 +217,7 @@ class MuJoCoFloatingPlatform:
         self.reset(initial_position=initial_position, initial_orientation=initial_orientation)
 
         while self.duration > self.data.time:
-            state = self.updateState() # Updates the state of the simulation.
+            state = self.getObs() # Updates the state of the simulation.
             # Get the actions from the controller
             action = model.getAction(state)
             # Plays only once every self.inv_play_rate steps.
@@ -451,7 +226,9 @@ class MuJoCoFloatingPlatform:
                 mujoco.mj_step(self.model, self.data)
                 self.updateLoggers()
 
-    def runLoopForNSteps(self, model, initial_position: np.ndarray=[0,0,0], initial_orientation:np.ndarray=[1,0,0,0], max_steps=502) -> None:
+        return self.logs
+
+    def runLoopForNSteps(self, model, initial_position: np.ndarray=[0,0,0], initial_orientation:np.ndarray=[1,0,0,0], max_steps=502) -> Dict[str, np.ndarray]:
         """
         Runs the simulation loop.
         model: the agent.
@@ -460,7 +237,7 @@ class MuJoCoFloatingPlatform:
         self.resetPosition(initial_position=initial_position, initial_orientation=initial_orientation) # Resets the position of the body.
         i = 0
         while i < max_steps:
-            state = self.updateState() # Updates the state of the simulation.
+            state = self.getObs() # Updates the state of the simulation.
             # Get the actions from the controller
             action = model.getAction(state)
             # Plays only once every self.inv_play_rate steps.
@@ -469,6 +246,8 @@ class MuJoCoFloatingPlatform:
                 mujoco.mj_step(self.model, self.data)
                 self.updateLoggers()
             i += 1
+        
+        return self.logs
     
     def plotSimulation(self, dpi:int = 120, width:int = 600, height:int = 800, save:bool = False, save_dir:str = "mujoco_experiment") -> None:
         """
